pyproject.toml
[project]
name = "stock-analysis-platform"
version = "0.1.0"
description = "Multi-agent stock analysis platform"
requires-python = ">=3.11"
dependencies = [
    "prefect>=2.14.0",
    "anthropic>=0.18.0",
    "akshare>=1.12.0",
    "tushare>=1.4.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "pydantic>=2.5.0",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
